update: extract Sidebar into standalone component and refine logout hover styling #44
